feat: Show a message when the job list is empty
Fixes #11. The Jobs page would previously be completely empty (besides
the heading) if there aren't any jobs. We now show a message there to
inform the user that the emptiness is intentional, and that the page
will refresh automatically once jobs arrive.

Error handling is also added to the page.