feat: Verify job ownership
In case the Renovate CronJob shares its namespace with other
resources, we need to ensure that any jobs we return are actually
related to the Renovate CronJob and not part of something else.

In the case of scheduled jobs, this is easily verified using the job's
owner references. In the case of jobs triggered manually via Foreman,
it appears we need to introduce a custom label by which we can then
filter (which is what this patch does).

When testing this patch, non-related jobs no longer show up in
Foreman while scheduled and manual jobs still do.