feat: Handle CronJob not found
Before this patch, if the Renovate CronJob could not be found, we
would return a 500 Internal Server Error and the UI would look
broken. Now, we return a 404 Not Found, log a warning, and let the
frontend show an error message component. The info cards are
updated to properly handle this case.