feat: Handle empty repository autodiscovery
When Renovate's autodiscovery turns up empty, for example when
triggering a manual run with a non-existing repository scope, the log
will not contain a list of matched repositories (not even an empty one).
However, once the pod has completed, we know for a fact that no
repositories are going to be discovered any more.
We can then show an empty list indicator, instead of continuing to
show a loading indicator,