fix: Skip parsing irrelevant log lines for progress
This patch reduces the time spent extracting progress by 1.5-10x
according to testing, depending on the number of repositories used
and whether debug or even trace logging was active. For trace logging
with a single repository, the time was reduced from >120ms to 18ms.
No situation could be found where the performance was degraded
due to the additional checks - it seems like a safe thing to do.