feat: Handle termination signals
This patch adds handling of SIGINT and SIGTERM so that the server
is closed gracefully. It also moves the code that starts the server into
a function (in `server.ts`), separate from the command-line entrypoint
(in newly-added `main.ts`). This way, each code part's responsibilities
are clearer than if the process exit logic was conflated with running
Fastify.