test: Add integration tests for local login
This patch tests the strategies, me, local login, and logout routes with
our integration test suite. While adding these tests, some non-optimal
server responses were discovered (such as returning errors as text
instead of JSON, and using SameSite=Lax cookies instead of Strict),
but these are completely non-critical and are annotated with TODO
for now, to be fixed in a follow-up code change.