fix: Ensure proper OIDC redirect_uri is used
openid-client manipulates the redirect URI we gave
based on the request URL, which will be wrong when
behind a reverse proxy. This is a workaround for
that based on upstream discussions.

Alternatively, it may be possible to configure the
trust proxy setting, but this is a burden on users
and not really necessary when we already have the
URL and could use it as-is.